Finished Freight auto for right pos
I finished the auto for red freight if the shipping element is in left pos and got the arm and winch positions roughly setup to where they will need to be for the other positions on a sticky note.
Also broke the claw by running the robot off a table accidently, should be easy to reprint it.
*Messed with strafePureEncoder adding in a driveTimeInSeconds argument that doesn't work currently.